Merget med main og lagt inn manglende tester for TilgangService
